Switch constructors with 1 arg as explicit (avoid implicit casting errors).
Cppcheck reported errors like:
"Class XXX has a constructor with 1 argument that is not explicit."

Indeed if there is only one argument in the constructor, the compiler is
allowed to implicitely cast the argument if it is used as another type.
Making the constructor as "explicit" force the compiler to check the type
of the argument when it is used.